requirements.txt
torchvision
Pillow
requests